local/local_requirements.txt
flask
opencv-python
qrcode
Pillow
